modularized UI controller
modularized UI controller
