+++ Assets/scripts/ScriptManager.cs
@@ -7,7 +7,6 @@ using TMPro;
 using System;
 using UnityEngine.Events;
 using UnityEngine.UIElements;
-using System.IO;
 using UnityEngine.SceneManagement;
 using MoonSharp.Interpreter.Debugging;
 using Unity.VisualScripting;
@@ -17,20 +16,14 @@ using UnityEngine.SocialPlatforms;
 
 public class ScriptManager : MonoBehaviour
 {
-    // UI elements
-    public GameObject inputField;
-    public TMP_InputField userInputField;
-    public TextMeshProUGUI userOutTextForDebug;
-    public TextMeshProUGUI userOutText;
-    public TextMeshProUGUI userOutTextFunctionDispaly;
+    // Declare the uIManager field
+    private UIManager uIManager;
 
     // The cube object
     [SerializeField]
     private GameObject cube;
     private Renderer cubeRenderer;
-    public GameObject ScreenButton;
-    public GameObject HelpWindow;
-    
+
 
     // Variables for the cube's color and size
     private Color newCubeColor;
@@ -45,8 +38,7 @@ public class ScriptManager : MonoBehaviour
     UnityEvent unityEvent = new UnityEvent();
     // Create a UnityAction object that represents a method
 
-    // The name of the current scene
-    private string sceneName;
+
     public int score = 0;
     float time = 0.0001f;
     float defaultTime = 1f;
@@ -60,15 +52,14 @@ public class ScriptManager : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
+        uIManager = GameObject.Find("UIManager").GetComponent<UIManager>();
+
         GameObject newCube = GameObject.Find("cube");
         //renderer = cube:GetComponent("Renderer")
         // Get the renderer component of the cube
         cubeRenderer = cube.GetComponent<Renderer>();
 
-        // Get the name of the current scene
-        sceneName = SceneManager.GetActiveScene().name;
-        sceneName += ".txt";
-
+        
 
         // Add a Rigidbody component to the cube game object and set its useGravity property to true
         Rigidbody rb = cube.GetComponent<Rigidbody>();
@@ -78,7 +69,7 @@ public class ScriptManager : MonoBehaviour
 
         if (data != null || data != "")
         {
-            userInputField.text = data;
+            uIManager.userInputField.text = data;
             
         }
 
@@ -87,6 +78,7 @@ public class ScriptManager : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
+        
         // Invoke any tasks that are currently registered with the UnityEvent
         unityEvent.Invoke();
 
@@ -95,7 +87,7 @@ public class ScriptManager : MonoBehaviour
         {
             // Display a message in the debug log
             Debug.Log("You are dead!");
-            RestartScene();
+            uIManager.RestartScene();
         }
     }
     // Set up a function that takes a string as an argument
@@ -212,23 +204,14 @@ public class ScriptManager : MonoBehaviour
         return position;
     }
     
-    public void RestartScene()
-    {
-        print("Level Restarted");
-        DataInputHoldingData.instance.dataInput = userInputField.text;
-        // Get the current scene name
-        string sceneName = SceneManager.GetActiveScene().name;
 
-        // Load the scene with the given name
-        SceneManager.LoadScene(sceneName);
-    }
     // This method loops through the list of tasks and waits the specified amount of time before executing each one.
     private IEnumerator DoTask()
     {
         for (int i = 0; i < listOfTasks.Count; i++)
         {
             // Update the userOutTextFunctionDispaly variable to show the name of the function being executed
-            userOutTextFunctionDispaly.text = listOfTasks[i].Method.Name;
+            uIManager.userOutTextFunctionDispaly.text = listOfTasks[i].Method.Name;
 
             // Add the task to the UnityEvent and wait the specified amount of time.
             unityEvent.AddListener(listOfTasks[i]);
@@ -255,12 +238,12 @@ public class ScriptManager : MonoBehaviour
         try
         {
             // Get the script text from the input field
-            string script = inputField.GetComponent<TMP_InputField>().text;
+            string script = uIManager.inputField.GetComponent<TMP_InputField>().text;
             Debug.Log("script: " + script);
 
 
             StartLua(script);
-            userOutTextForDebug.text = "None Error messages from Lua";
+            uIManager.userOutTextForDebug.text = "None Error messages from Lua";
 
             
             if (listOfTasks.Count == listOfTime.Count)
@@ -278,19 +261,19 @@ public class ScriptManager : MonoBehaviour
         {
             // if a syntax error was detected, display an error message to the user
             Debug.Log("Syntax error: " + ex.Message);
-            userOutTextForDebug.text = "Syntax error: " + ex.Message;
+            uIManager.userOutTextForDebug.text = "Syntax error: " + ex.Message;
         }
         catch (ScriptRuntimeException ex)
         {
             // if a runtime error was detected, display an error message to the user
             Debug.Log("Runtime error: " + ex.DecoratedMessage);
-            userOutTextForDebug.text = "Runtime error: " + ex.DecoratedMessage;
+            uIManager.userOutTextForDebug.text = "Runtime error: " + ex.DecoratedMessage;
         }
         catch (Exception ex)
         {
             // if any other exception was thrown, display the error message to the user
             Debug.Log("Error: " + ex.Message);
-            userOutTextForDebug.text = "Error: " + ex.Message;
+            uIManager.userOutTextForDebug.text = "Error: " + ex.Message;
         }
 
         ResetCubeData();
@@ -309,7 +292,7 @@ public class ScriptManager : MonoBehaviour
         //randomChannelTwo = g;
         //randomChannelThree = b;
 
-        UserOutTextFunctionDispaly("SetCubeColor");
+        uIManager.UserOutTextFunctionDispaly("SetCubeColor");
 
         listOfTasks.Add(() => { });
         listOfTime.Add(time);
@@ -317,7 +300,7 @@ public class ScriptManager : MonoBehaviour
     public void SetCubeSize(float size)
     {
 
-        UserOutTextFunctionDispaly("SetCubeSize");
+        uIManager.UserOutTextFunctionDispaly("SetCubeSize");
         // set the scale of the cube to the specified size
         //cube.transform.localScale = new Vector3(size, size, size);
 
@@ -326,35 +309,35 @@ public class ScriptManager : MonoBehaviour
     }
     public void MoveForward(float Time = 1f)
     {
-        UserOutTextFunctionDispaly("MoveForward");
+        uIManager.UserOutTextFunctionDispaly("MoveForward");
         print($"MoveForward {Time}");
         listOfTasks.Add(MoveF);
         listOfTime.Add(Time);
     }
     private void MoveRight(float Time = 1f)
     {
-        UserOutTextFunctionDispaly("MoveRight");
+        uIManager.UserOutTextFunctionDispaly("MoveRight");
         print($"MoveRight {Time}");
         listOfTasks.Add(MoveR);
         listOfTime.Add(Time);
     }
     private void MoveLeft(float Time = 1f)
     {
-        UserOutTextFunctionDispaly("MoveLeft");
+        uIManager.UserOutTextFunctionDispaly("MoveLeft");
         print($"MoveLeft {Time}");
         listOfTasks.Add(MoveL);
         listOfTime.Add(Time);
     }
     private void MoveBack(float Time = 1f)
     {
-        UserOutTextFunctionDispaly("MoveBack");
+        uIManager.UserOutTextFunctionDispaly("MoveBack");
         print($"MoveBack {Time}");
         listOfTasks.Add(MoveB);
         listOfTime.Add(Time);
     }
     private void Wait(float Time = 1f)
     {
-        UserOutTextFunctionDispaly("Wait");
+        uIManager.UserOutTextFunctionDispaly("Wait");
         print($"Wait {Time}");
         listOfTasks.Add(() => { });
         listOfTime.Add(Time);
@@ -390,7 +373,7 @@ public class ScriptManager : MonoBehaviour
         {
             displacement = new Vector3(0, 0, -(float)distance * speed * Time.deltaTime);
         }
-        UserOutTextFunctionDispaly("Move");
+        uIManager.UserOutTextFunctionDispaly("Move");
         // Add a task to the list of tasks that moves the cube by the specified amount in the specified direction.
         // The task will wait for the specified delay plus the time delta before executing.
         listOfTasks.Add(() => cube.transform.position += displacement);
@@ -431,7 +414,7 @@ public class ScriptManager : MonoBehaviour
     public void Teleport(float x, float y, float z)
     {
 
-        UserOutTextFunctionDispaly("Teleport");
+        uIManager.UserOutTextFunctionDispaly("Teleport");
         // Get the current position of the game object
         Vector3 currentPosition = transform.position;
 
@@ -448,7 +431,7 @@ public class ScriptManager : MonoBehaviour
     // This method sets the speed at which the cube moves.
     private void SetCubeSpeed(float speed)
     {
-        UserOutTextFunctionDispaly("SetCubeSpeed");
+        uIManager.UserOutTextFunctionDispaly("SetCubeSpeed");
         // Set the speed field to the specified value
         //this.speed = speed;
         listOfTasks.Add(() => this.speed = speed);
@@ -461,7 +444,8 @@ public class ScriptManager : MonoBehaviour
         Script lua = new Script();
 
         // Register the "print" function so that the script can print messages to the debug log
-        lua.Globals["print"] = (Action<DynValue>)PrintToDebugLogAndTextArea; ;
+
+        lua.Globals["print"] = (Action<DynValue>)uIManager.PrintToDebugLogAndTextArea;
 
         // Register the custom functions that the script can call
         RegisterFunctions(lua);
@@ -488,38 +472,7 @@ public class ScriptManager : MonoBehaviour
         //dynValue = DynValue.NewBoolean(somethingInRight);
         //lua.Globals.Set("somethingInRight", dynValue);
     }
-    //    -- Access the values of the global variables in the Lua script
-    //local somethingInFront = somethingInFront
-    //local somethingInBack = somethingInBack
-    //local somethingInLeft = somethingInLeft
-    //local somethingInRight = somethingInRight
 
-    //-- Check if something is in front of the cube
-    //if somethingInFront then
-    //    -- Do something
-    //end
-
-    //-- Check if something is behind the cube
-    //if somethingInBack then
-    //    -- Do something
-    //end
-
-    //-- Check if something is to the left of the cube
-    //if somethingInLeft then
-    //    -- Do something
-    //end
-
-    //-- Check if something is to the right of the cube
-    //if somethingInRight then
-    //    -- Do something
-    //end
-
-    //Crashes the program
-//    local somethingInFront = somethingInFront
-
-//while not somethingInFront do
-//    MoveForward()
-//end
 
 //print("Something is in front of the cube!")
     // This method registers the custom functions that the script can call.
@@ -570,28 +523,7 @@ SetCubeColor(5,5,5)
         */
     }
     // This method prints the type and value of the specified value to the debug log and the Text object
-    private void PrintToDebugLogAndTextArea(DynValue value)
-    {
-        // Get the type of the value
-        string type = value.Type.ToString();
-
-        // Convert the value to a string
-        string message = value.ToObject().ToString();
-
-        // Print the type and value to the debug log
-        Debug.Log("Printing " + type + " value from Lua script: " + message);
-
-        // Set the text of the Text object to the type and value
-        userOutText.SetText(type + ": " + message);
-    }
-    public void UserOutTextFunctionDispaly(string text)
-    {
-        //userOutTextFunctionDispaly.text = text;
-    }
-    public void ResetText()
-    {
-        userInputField.text = "";
-    }
+    
 
     public void Stop()
     {
@@ -614,50 +546,5 @@ SetCubeColor(5,5,5)
         cube.transform.localScale = Vector3.one;
 
         
-    }
-
-    public void SaveInput()
-    {
-
-        // get the user input
-        string userInput = inputField.GetComponent<TMP_InputField>().text;
-
-        // write the user input to a file
-        File.WriteAllText(sceneName, userInput);
-    }
-
-    public void LoadSave()
-    {
-        // read the user input from the file
-        string userInput = File.ReadAllText(sceneName);
-
-        // set the input field to the saved user input
-        inputField.GetComponent<TMP_InputField>().text = userInput;
-    }
-
-    public void HideOrViewConsole()
-    {
-        if (ScreenButton.activeSelf == true)
-        {
-            ScreenButton.SetActive(false);
-
-        }
-        else
-        {
-            ScreenButton.SetActive(true);
-        }
-    }
-
-    public void HideOrViewHelpWindow()
-    {
-        if (HelpWindow.activeSelf == true)
-        {
-            HelpWindow.SetActive(false);
-
-        }
-        else
-        {
-            HelpWindow.SetActive(true);
-        }
     }
 }
